+++ lending/examples/companiesUpdate.example.ts
@@ -8,7 +8,7 @@ dotenv.config();
  * Example usage of the @speakeasy-sdks/lending SDK
  *
  * To run this example from the examples directory:
- * npm run build && npx tsx companiesUpdate.ts
+ * npm run build && npx tsx companiesUpdate.example.ts
  */
 
 import { LendingTs } from "@speakeasy-sdks/lending";